examples/docker/video-rag/requirements.txt
haystack-ai
elasticsearch-haystack==0.1.2
solara
openai
python-dotenv
pydub
moviepy
modal